Update README.md
Bubble Shooter is a basic arcade game where player shoots colored bubbles into a level that consists of other colored bubbles. When the bubble of the player collides with the bubbles in the level and a cluster of three or more bubbles with the same color is formed, the cluster is removed and the player is rewarded points. This game is build on javascript and it is a variation of match-3 mechanic. When a cluster of three or  same color created it break down. Then, the player is given the next bubble and can shoot again. If any of the bubbles reaches the bottom of the level, the player has lost the game.